lots of changes to evaluation and switching code

+++ ramp_planner/src/population.h
@@ -23,7 +23,7 @@ class Population {
     const RampTrajectory  get(const unsigned int i) const;
     const int             getIndexFromId(const uint16_t id) const;
     const std::vector<RampTrajectory> getTrajectories() const;
-    const int             getMinFitness() const;
+    const double          getMinFitness() const;
     const bool            contains(const RampTrajectory rt) const;
     const bool            feasibleExists() const;
     const bool            infeasibleExists() const;
